.github/scripts/analytics/ydb_metrics_analytics_system/mas_ui/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

